refactor:  Nav Tabs component로 분리
